refactor(clientes): address A3 review — drop unused clock/zona_nombre, add worker tests
- Remove unused `clock outbound.Clock` field+param from DirectoryReconcileWorker;
  tick already calls time.Now() directly. Drop clock arg from wiring call site.
- Remove dead ZonaNombre field from outbound.DirectorioDoc and its assignment in
  reconciliar_directorio.go; field was never mapped into ClienteDoc (silently dropped).
- Add lifecycle test suite for DirectoryReconcileWorker mirroring ReindexWorker:
  StartStop, StartIdempotent, StopWithoutStart, WarmupFailsDegradesSafely.
- Change ClienteDoc.Score from float64 to int (matches DirectorioDoc.Score int,
  range 0-100); drop the float64() cast in mapDoc, update index_test assertions.
- Replace time.Sleep(3s) in integration test with a 30s poll-with-timeout
  against the Meilisearch stats endpoint (500ms interval).

+++ internal/clientes/infra/clientessearch/index.go
@@ -77,7 +77,7 @@ func mapDoc(d outbound.DirectorioDoc) ClienteDoc {
 		ConSaldo:           d.ConSaldo,
 		Segmento:           d.Segmento,
 		EstadoPago:         d.EstadoPago,
-		Score:              float64(d.Score),
+		Score:              d.Score,
 		RecenciaDias:       d.RecenciaDias,
 		Estatus:            d.Estatus,
 		SegmentoOrden:      segmentoOrdinal(d.Segmento),
